[gerrit fetch] Improve Gerrit fetch regex
Handle quoted URLs, as well as "checkout -b" (ignoring the suggested
branch name, though) and "cherry-pick".

If the initial setting is not showing the "create branch" (or tag)
section, add an empty placeholder label so that the dialog has the
right size initially and can show all sections properly when the user
then toggles sections.

Change-Id: Ie5a0f3ca9821e100d5dac334c9fc4995c6a46263
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>