Refactor the FetchGerritChangePage
Move most of the logic into an abstract super class. Leave only
logic relating to fetching change refs and dealing with change
numbers in the concrete subclass.

This opens the door to using this dialog also for fetching Github
pull requests, or Gitlab merge requests, or similar remote feature
branches.

Bug: 544815
Change-Id: Ic110ff49754eea0b201883b5f5a1e9b1a06b5d39
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>