[releng] Eliminate JBoss repository-utils
The JBoss repository-utils plug-in is unavailable as of 2021-10-04;
the JBoss Nexus causes connection resets. Eliminate this plug-in from
our build.

The <associateSites> feature is replaced by tycho standard
<repository-reference>s in category.xml. (Available since tycho 2.4.0.)

The web site for the repository is built now explicitly using standard
maven plug-ins. XSLT is used to transform the category.xml generated by
tycho into a HTML fragment, from which the index.html is then assembled.

Change-Id: Ib88e720f7ea9a9758ac4fef85766bd77b7003372
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>