[fetch pr] Generalize fetching pull requests
Refactor the wizard for fetching a pull request such that it is no
longer specific to Github. Introduce a GitServer abstraction that
encapsulates all UI labels and everything having to do with IDs of
pull requests.

This change does not yet add other git server types; it only prepares
for doing so.

Change-Id: I99b7432224a3c7eeb6adb4faacd48caf5fb9a9da
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>