[repo view] Ensure the correct commits are compared for Refs
The repositories view contains a rather ugly hack: it compares RefNodes
only by name. So even if the commit a Ref points to changes, the Common
Navigator still sees the same RefNode object and doesn't update its
internal selection cache.

The decorator for these nodes does update, so the UI reflects the
correct state. But the Common Navigator's selection cache doesn't.

This was done to avoid a nasty flicker in the repositories view every
time a branch moves. But it also means that code cannot rely on the
object ID obtained from a RefNode's Ref since that Ref instance might
be outdated. Instead, code has to resolve the Ref afresh. (The name of
the Ref is reliable.)

Change the CompareCommand of the repository view accordingly.

Since this is so arcane and wasn't documented anywhere except in a
brief mention of this flicker in the in bug 546455 and in the commit
message of commit 9244bc84, add comments to RepositoryTreeNode.equals()
explaining all this.

Bug: 576959
Change-Id: I0af5a32958501cef7dfe663b7bcf077b0860698d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>