Add a preference for the JGit text buffer size
Add a new setting to the Window Cache preference page where the user
can change this "text buffer size". It's the number of bytes that JGit
examines to determine whether a file or blob is binary or text, and if
text, whether it has CR-LF line endings. The preference can be set to
values in the range of 8 to 128KiB.

The JGit default is a mere 8kB. Set a larger default of 32KiB.

Bug: 575393
JGit-Dependency: I86e4f5f62e14f99b35726b198ba3bbf1669418d9
Change-Id: I4ed911abbe638dc723e19770e641653532dd12b5
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>