Update patch set 4
Patch Set 4: Verified+1 Code-Review+2

Overruling the failing test which is disabled in the next change
it is only failing on the CI and works locally.

We need to tag RC1 hence overrule this failure to unblock creating this milestone build.
Will reenable the test as soon as we found the reason for the failure on the CI.

Patch-set: 4
Reviewer: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1
Label: Code-Review=+2