Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 2:
> 
> Also: how does this interact with the ICommitMessageProvider/ICommitMessageProvider2 extension point mechanism? Which one takes precedence, and which one _should_ take precedence?

TBH, I did not know those classes. Nevertheless, the current flow is that the StagingView sets the message content (if available from the template) during loadInitialState() and then commitMessageComponent.updateUI() triggers calculateCommitMessage(commitMessageWithCaretPosition) which takes a shortcut if commitMessage is already filled. So in that case the prefilled message takes precedence.

Since I do not know the functioning of ICommitMessageProvider/2 I cannot answer if that is the way it _should_ be.

Patch-set: 3