Fix history job for empty histories
Correct the quick fix done in commit 0a4bc695 for ensuring that the
the commit graph table is cleared when there are no commits at all
loaded.

The != 0 condition led to the history of a non-committed file being
loaded ad nauseam. So remove that condition, but initialize the field
to -1 to ensure that on the very first time we load such an empty
history, we do update the UI.

Bug: 577853
Change-Id: I6a1a5c379caf22684ba39b4ad505aca7374912a4
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>