[fetch pr] Simplify tooltip message construction
Simplify the way the tooltip message is constructed. Formerly we used
a quoted placeholder '{0}', which might be confusing for translators.

Make the message use two plain placeholders {0} and {1} instead of
'{0}' and {0}. To preserve behavior, replace {0} by "{0}", so that
the second replacement inside UIUtils.addContentProposalToText() still
has its expected placeholder {0} in the message.

Bug: 577764
Change-Id: I62c98b2d94d1056445e76330e93237c9937884b3
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>