Update patch set 1
Patch Set 1:

Thank you, Gayan, for this contribution.

I find this highly confusing.

First, I believe showing untracked files should be the default; the logic needs to be switched around. I.e., a button with a red diagonal across that file icon, and text "Hide untracked files".

Second, the setting should be persisted, possibly even per repository, so that if the user closes and re-opens the view, it is still effective.

Third, what if the user selects an untracked file in the package explorer and chooses "Team->Commit..." from the context menu and has "Use staging view for committing" enabled (which is the default)? In this case, I think untracked files should be shown always, since it is highly confusing and error prone not to show them. Too easy to forget to commit a new file.

Fourth, the "Unstaged Changes" header shows the number of changed files, for instance "Unstaged Changes (1)". When filtering the view, for instance by "*.vvv" (where no matching changed file exists), it'll show "(0/1)", i.e., the user has _some_ indication that not everything is being shown. (Personally, I think it's not prominent enough, but at least there is some feedback.) That must also be the case when this "hide" option is on. (And seeing that the button is checked is not sufficient, since it might be checked, but there are no untracked files.)

I'm not convinced this is a good idea in general. Why do you want to hide untracked files at all?

Finally a nitpick: please follow the EGit/JGit commit message guidelines (bug reference in the footers). See [1].

[1] https://wiki.eclipse.org/EGit/Contributor_Guide#Commit_message_guidelines

Patch-set: 1
CC: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>