Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thank you, Gayan, for this contribution.
> 
> I find this highly confusing.
> 
> First, I believe showing untracked files should be the default; the logic needs to be switched around. I.e., a button with a red diagonal across that file icon, and text "Hide untracked files"

> 
> Second, the setting should be persisted, possibly even per repository, so that if the user closes and re-opens the view, it is still effective.
> 
What if we make this toggle just a toggle which is by default showing the untrack files ? it will not listen to settings, As you point out it could be a problem when solving conflicts as well.

> Third, what if the user selects an untracked file in the package explorer and chooses "Team->Commit..." from the context menu and has "Use staging view for committing" enabled (which is the default)? In this case, I think untracked files should be shown always, since it is highly confusing and error prone not to show them. Too easy to forget to commit a new file.

Will this be solved if we use the above mentioned strategy ?

> 
> Fourth, the "Unstaged Changes" header shows the number of changed files, for instance "Unstaged Changes (1)". When filtering the view, for instance by "*.vvv" (where no matching changed file exists), it'll show "(0/1)", i.e., the user has _some_ indication that not everything is being shown. (Personally, I think it's not prominent enough, but at least there is some feedback.) That must also be the case when this "hide" option is on. (And seeing that the button is checked is not sufficient, since it might be checked, but there are no untracked files.)

Will this be solved if we use the above mentioned strategy ? because we will treat this as a filter and may be we can change the label for (0 of 1).

> 
> I'm not convinced this is a good idea in general. Why do you want to hide untracked files at all?
> 
The reason i thought it will be a good idea is that, i really like the stage view since it doesn't block the UI like the commit dialog. in my workflows sometime i get find files which i don't want to commit, so having the untrack files option like in the commit dialog helps to have more focus view. My requirement is to have the commit dialog UI in the view instead of a dialog :). WDYT ?


> Finally a nitpick: please follow the EGit/JGit commit message guidelines (bug reference in the footers). See [1].
> 
> [1] https://wiki.eclipse.org/EGit/Contributor_Guide#Commit_message_guidelines
Sure i can correct it in my next push.

Patch-set: 1