Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > Second, the setting should be persisted, possibly even per repository, so that if the user closes and re-opens the view, it is still effective.
> > > 
> > What if we make this toggle just a toggle which is by default showing the untrack files ? it will not listen to settings, As you point out it could be a problem when solving conflicts as well.
> > 
> You mean, basically just set the button checked initially? We could try that.

I totally agree that we should not make any workflow changes to current users who are familiar with the state view. I really appreciate your thoughts which really helped me to see things i didn't thought of.

Yes checked the button by default regardless the mode of the staging view and also the user choice is not persisted at all and it will not listen to any of the settings either. So when this change is received by a user it will be basically transparent for them functionally, unless they try to use it. To make it more closer to current functionality, when every the content of the view is updated by 
- user selecting to commit
- conflict on a merge or rebase
- ammend commit

we will reset the button state to default. But at the same time if your commit file list is cluttered with lot of untracked files which you don't want to commit, then you can choose to hide them, and its a choice that the user make. To remind user there are untracked files we can update the label.

May be in the long run we can remove the commit dialog if that is not used by lot of users if the functionality of the dialog is covered in the view.

Patch-set: 1