Clean commit messages consistently, and syntax-color comment lines
Implement proper commit message cleaning according to git config
"commit.cleanup". With the default git config, this changes the
behavior of EGit: it now _does_ remove comment lines in commit
messages by default. Previously, it didn't, except for squashing,
where JGit did it unconditionally.

Cleaning commit messages entered by the user in the UI is more
consistent with the way C git behaves: it cleans messages obtained
via an editor, but does not clean messages given on the command-line
as in "git commit -m 'Some message'".

To make the user aware of this new behavior, provide syntax coloring
for comment lines in commit messages in the SpellcheckableMessageArea.
Such syntax coloring makes only sense if the lines are indeed removed
the message upon commit, so the implementation is only active for
clean-up mode "strip".[1] For read-only display of commit messages,
no syntax coloring for comment lines is done.

The color used is configurable via the Eclipse preferences and is
themeable. The default colors for the light and dark themes correspond
to the default syntax coloring configuration of JDT for Java single-
line comments.

Since the staging view re-uses the same SpellcheckableMessageArea for
multiple repositories, and the clean-up mode may be different between
repositories, provide operations to change the clean-up mode of an
existing SpellcheckableMessageArea and update the presentation.

Change the interactive rebase InteractiveHandlers to the new interface
InteractiveHandler2 to ensure this syntax coloring also works for
rewording or squashing.

There is no support (yet) for automatically updating the presentation
in the SpellcheckableMessageArea when the git config changes. The
view's refresh action does update the presentation, though. Switching
between repositories also does so.

Bug: 553065
Bug: 578173
JGit-Dependency: Ia0040c247884e684587dd45d6cb85f8b72a4b876
Change-Id: I07ceedcc781b50bde63805ff0bef9fde06902c80
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>