Update patch set 4
Patch Set 4:

Works better. The setting is reset to "show untracked" on every repository switch, which I find fine. It also resets on every index change (creating a new file, changing an existing file, unstaging a staged file). Is that intended?

Also toggling the setting always expands the tree. That's a bit strange.

When there are no untracked files, but a changed file, and then the setting is toggled off, it shows "Unstaged Changes (1/1)". I think that's OK.

Patch-set: 4