Add .gitattributes for OSGI-INF files
The XML files for the OSGi declarative services in directory OSGI-INF
are automatically generated from the Java annotations in the sources
during a build. If the file exists and was checked out on Windows with
CR-LF line endings, the generator may produce mixed line endings when
it updates the file (for instance, on a full build on Windows). Git and
JGit then consider the file changed, even if there is no real change.

The mixed line endings on Windows come from the PDE DS annotation
processor (org.eclipse.pde.ds.internal.annotations.AnnotationVisitor),
which normalizes properties with multiple values to an element body
with each value separated by '\n'.

So add .gitattributes specifying "text eol=lf" for these files.

Change-Id: Ic50963757e9b8b289c7392842f291dec383e6cdd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>