[dark mode] Use default background for tables and trees
Use the default background #2F2F2F instead of the somewhat lighter
background the default Eclipse dark theme assigns to tables and trees.

The trees in other viewers also use this background; there is no reason
to use a lighter shade of gray. With this change, all elements in the
git history and git staging views use the same background, and it's the
same background as is used in other views.

Bug: 578611
Change-Id: Ic69c8c590d0cf78d46eaaed13e62b55ae8b9309f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>