[dark mode] No CSS styling for the commit message in the staging view
The commit message editor in the staging view behaves like a text
editor and uses the text editor preferences for colors and fonts.
Unfortunately, the text editor color preferences are overwritten
by CSS styling, with the result that they are effectively ignored
in dark mode. But if a user defines non-standard colors, these choices
should be respected no matter what the CSS defines. The CSS should only
define the default appearance if not overridden by user-settable
preferences.

The fix is to disable CSS styling of the StyledText widget. The
SpellcheckableMessageArea is implemented like a text editor, updating
correctly when the editor preferences change. Themeing updates these
preferences, and thus the net result is that the commit message editor
still honors the theme, plus additionally any overrides the user might
have defined via the preferences.

Bug: 559321
Change-Id: I88e4b1d8d2653aec275241a78064a7e2511a0380
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>