[staging] Improve "Commit & Push" push behavior
"Commit & Push" in the staging view pushed only specifying the remote.
This would push as if "push.default=current" was specified if the remote
had no push refspecs, but if it had, it might even push several branches
or totally unrelated branches.

Make sure that at most the current branch is pushed. Take into account
the branch configuration: implement a cross between push.default
"upstream" and "current": if an upstream branch is specified, use it,
otherwise fall back to pushing to a branch with the same name as the
local branch.

For the current EGit/JGit implementation, this gives a more intuitive
push behavior. It ignores the git config for "push.default", though.
(The default from the git config would be "simple", and would forbid
the push if no upstream branch was configured, or it had a name
different from the local branch name.) This is by design; using "Commit
& Push" should never use push.default=matching, even if the git config
says so.

Bug: 441031
Change-Id: I61f4fa370ed9c39753b9de52f84c986f46ca2bd1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>