Respect the push.default git config in "Push to Upstream"
Let the command determine how and what exactly to push from the git
config.

Move the remote determination from SimpleConfigurePushDialog to
PushOperation. It is not UI specific, and moving it enables its use
in the gitflow CurrentBranchPublishOperation. Rewrite that operation
to always use PushDefault.CURRENT.

"Push to Upstream" uses whatever is configured in the git config. This
may in some cases fail to push, for instance if the branch has an
upstream with a different name configured, and push.default is
"simple" (the default). In bug 431872 it was suggested to ask the user
up-front in this case what should be done. If so, then that should also
be done in "Commit & Push" in the staging view.

I also wonder if the user should be warned if "Push to Upstream" would
push several branches (if the remote has a push refspec with wildcards,
or push.default is "matching").

Bug: 431872
Change-Id: If9c30553f2b217a054e71d8b6c27b34125f8edc1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>