[repo view] Open branches, tags, or stashes in the commit viewer
Add a command handler for opening branches, tags pointing to commits, or
stashed commits in the commit viewer. Unify the command/handler setup in
plugin.xml so that there is only one command for this, with different
handlers active for different kinds of selections.

Change the menu entry for stashes to use the new handler; it was already
wired to use the generic "Open" handler used otherwise for
double-clicks, but that one works only for single selections.

As in the history view, the new handler is active for multi-selections
if all selected items are branches, tags, or stashes, even from
different repositories. It is not active if the selection includes any
other kind of node, or if a single tag node is selected but the tag
doesn't point to a commit.

Bug: 578711
Change-Id: I90bd93c2cc6a2039c0651b3ccebf101b6ba89787
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>