Push to Upstream: use more git config, and ask user if needed
This is a bit more involved than just letting the underlying JGit
PushCommand work with the git config. JGit implements the canonical
C git behavior, which sometimes denies the push, for instance, if
push.default=simple (the default), and no upstream branch is configured
for the currently checked-out local branch. But exactly this case is
very common, and just denying the operation is not user-friendly. It's
much better in this case to simply open the "Push Branch..." wizard and
let the user decide what should be done.

Because of that EGit re-implements the logic already present in JGit
for figuring out what to push when no refspecs are given. EGit adds
its own twists to it:

* It opens the "Push Branch..." wizard in several cases where JGit or
  C git would simply deny the operation.
* It warns about pushing multiple branches (remote config with wildcard
  push refspecs, or push-default=matching).

This change also drops the requirement that "origin" is used only if it
has a push refspec. Pushing is possible without a push refspec. Note
that EGit auto-configures repositories cloned from Gerrit to have a
remote with a push refspec "HEAD:refs/for/master", so this should not
cause surprises for Gerrit users.

Bug: 431872
Change-Id: If9c30553f2b217a054e71d8b6c27b34125f8edc1
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>