PushBranchWizard: allow bypassing the confirmation page
This wizard produces a single RefSpec for pushing a single branch or
commit. The confirmation page doesn't add much value in this case in
typical simple workflows when a remote is already configured.

Allow the user to bypass the confirmation page initially and to finish
the wizard from the initial page directly. Once the confirmation page
has been shown, the wizard can be finished only from that page, though.
If the repository has no remotes yet, the wizard has an initial page to
create one. In that case, the confirmation page is still mandatory.

Bug: 577079
Change-Id: Ic82b7c08c8c8e47cf82f0b14147e93765415a454
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>