[staging] Flip the UI logic for hiding untracked files
Having a default-selected button in the toolbar of the unstaged viewer
is a strange UI. Normally, the user has to _select_ a button to effect
a change, not deselect it.

So flip the logic: instead of having a "Show untracked files" button
that is initially selected, let's have a "Hide untracked files" button
that is initially _not_ selected.

Change the tooltip, rename the action internally, and give it a new
icon: a struck-out untracked file icon.

Bug: 560748
Change-Id: I23920850565b6a8e48dc278601ca75b22c311de6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>