[history] Fix UI freezes and word-wrap in commit & diff viewer
Recomputing the size of the combined commit & diff container could
cause extensive UI freezes and other UI lagginess if the diff was long
because of several reasons:

1. The size was recomputed far too often. It was done every time the
   text presentation in either the commit or the diff viewer was
   invalidated. But even hovering over a hyperlink invalidates part
   of the text presentation.
2. There are two methods of computing the content size of a scrolled
   composite, and EGit used the one that is apparently somewhat slower.
3. Word-wrapping was on by default in these viewers, even though they
   didn't wrap because of a bug in the size calculation. Computing the
   rendered size of a text with word-wrapping is slower than without.

Fix this by only recomputing the size when the input changes, using the
apparently somewhat faster method of updating the scrolled composite,
and switching off word-wrapping by default. When word-wrapping is on,
set a fixed width and compute only the height; this makes word-wrapping
work.

Additionally resizing the viewer by dragging the containing sash was
super slow if the diff was large because EGit recomputed the size on
every ResizeEvent. SWT emits ResizeEvents continuously while the
control is being resized.

Switch off live re-flowing the text (through continuously recomputing
the size) while the user resizes the control. Only do this once after
the user has stopped resizing the control. Resizing the control with
word-wrap on may still occasionally lag, but much less than before.
With word-wrap off, there is no need to recompute sizes at all, so in
this case resizing is now perfectly smooth.

Bug: 430226
Change-Id: If6a598e0ecea437ee2c0b139371ccc148a1cb57d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>