Fix blocking during branch rename
The BranchRenameDialog must return successfully only once the branch
has been renamed, so just using a job is not feasible.

Blocking could occur because the RenameBranchOperation actually ran
as an IWorkspaceRunnable with a scheduling rule. Neither is necessary,
though, as this is pure repository operation.

Remove the workspace stuff and perform the git operation directly.
The git operation itself should run fairly quickly, and is not
cancellable once begun, so adding some progress monitor in the UI
is not necessary.

Bug: 552686
Change-Id: Ic35aa7fb00456a8f83c5ebf8d103ac9baf14794b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>