Speed up showing fetch results
The fetch result viewer shows branches updated, with the number of
commits fetched per branch in the label and the commits as children,
and the file diffs as children of the commits, and it auto-expanded to
level 2, i.e., showing the commits. Plus it used the guaranteed-unique
OID abbreviation.

All this required a lot of file accesses to read objects from the git
repository, which could block the UI for an extended time if many
commits were fetched.

Use the deferred loading mechanism to get all expensive operations off
the UI thread. Trigger an initial background load per branch to
determine the number of commits. Show the count only once it is
available. Also use deferred loading for the file diffs. Do not expand
all branches; only expand the first branch shown in the viewer.

The viewer also loaded the children of a branch in the ViewerComparator
in an attempt to order branches before tags (according to the inline
comment). This cannot be done anymore; instead simply sort the updates
by ref name.

Minor further changes:
* Use TreeViewer.setUseHashing(true).
* Use a BaseWorkBenchContentProvider; this doesn't need the IResource
  tracking the WorkbenchContentProvider does.
* Simplify creating the table in the PullResultDialog; give it fixed
  size hints as in the FetchResultDialog.

Bug: 579053
Change-Id: I16c3a53c4f80c23b3e56eb4eda49e78acacd0395
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>