Use core.commentChar when editing commit messages
Determine the comment character via the git config, or in the case of
merge messages after a conflict as resolved, from the "Conflicts:" line
in the merge message. (C git does not localize that text.) For mode
"auto", persist the character in the CommitMessageComponentState so
that it can be re-used from there when repositories are switched.

For rewording commit messages, JGit already determines the correct
comment character when it calls the edit callback handler.

Bug: 579292
JGit-Dependency: Idca19284a0240cd322e7512ea299a03658e1b2c1
Change-Id: I5e1a8e9fd9b5f7cfc35f8e728d5c1df89c6f9bff
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>