[squash] Ensure message is cleaned even if dialog is canceled
Return VERBATIM only if the dialog was not canceled. In that case,
the dialog already cleaned the message. But if it is canceled, return
the original CleanupMode and let JGit clean the message.

Note that a squash currently cannot be aborted. Canceling the dialog
only cancels editing the commit message. This is bug 451817.

Change-Id: If8e3f06787ae1322f0da0b8ebaa7347fd488bb0c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>