[staging] Push options for "Commit & Push"
Commit & Push performs the equivalent of

  git push <remote> HEAD:<upstream branch>

The upstream branch is determined from the branch.<name>.merge
configuration of the currently checked out branch; if none is set,
the upstream branch is assumed to have the same name as the current
local branch.

However, in some circumstances one wants to perform a force push in
this case. Add UI to set the force flag, so that "Commit & Push" will
do

  git push <remote> +HEAD:<upstream branch>

Because screen space is a bit tight if a user has a staging view with
vertical layout, for instance in the right part stack instead of at the
bottom, use a ToolItem with a drop-down menu for this flag. This needs
less space than a checkbox "[ ] Force Push". Because a menu with a
single item looks stupid, also put the "Always Show Push Dialog"
setting into that drop-down menu.

Both settings are persisted separately for each repository. The setting
for showing a dialog always is initialized from the global preference.
The "force push" setting is by default false, but will remain "true" for
any repository where the user switched it on.

When "Force Push" is checked, the icon of the ToolItem gets a little
"+" overlay.

This settings ToolItem is available only if EGit will not use the
"Push to Gerrit" dialog. That dialog is used if the remote has a
Gerrit-specific configuration, and the commit has a Change-Id line.
Force pushing a new patch set to Gerrit makes no sense, and the dialog
will be shown always anyway.

Bug: 411219
Change-Id: I5a66939ee2345db2cb721e3bb32d2849e48743be
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>