Introduce RepositoryCommitOpener to reduce duplicate code
There were three places in EGit UI that used identical code for opening
RepositoryCommits. Factor out this code.

The original code used OpenStrategy.activateOnOpen() instead of the
"activate" flag given. This is not necessary; open() is called only via
OpenAndLinkWithEditorHelper.InternalListener, which already passes in
the result of OpenStrategy.activateOnOpen().

Change-Id: I01b551132e9542130107e4b3b62a48555a342637
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>