Decoration Calculation may interrupt UI thread
If egit decorator is initialized from non UI thread, it shouldn't use
syncExec() because it doesn't know which code it may block - and IF it
blocks, UI task may be interrupted to avoid deadlocks and that could
lead to bad side effects.

So schedule color init on UI thread and proceed with other init without
blocking UI. In that case decorator will use all defaults until UI is
able to process the async events queue, which is OK. If the init happens
on UI thread, simply run the init code directly.

Bug: 579964
Change-Id: I61e72c7343bbe716614a4c4f7bd3362bb55b7090