Remove ToolsUtils.informUserAboutError()
This was added for the external difftool support, but is actually
unnecessary. The only place it was called was in an exception handler
in a background job. Let the Eclipse job framework deal with this by
returning an error IStatus instead.

Also move all CompareUtils_* externalized strings together, and use
placeholder substitution instead of string concatenation to make the
life of translators a little easier.

Change-Id: I6611a6b42c3519d844e2e09575c1340a9d87dd92
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>