[staging] Show core.commentChar in a tooltip
Add a tooltip to the "Commit Message" section text with a message
indicating what the current core.commentChar is. Especially when the
setting is "auto", it may not be clear to the user which character
could be used for comments. (Though in general a user is unlikely to
enter comment lines as they will be removed anyway, and will see
existing comments via syntax highlighting. Nevertheless I find it
marginally useful.)

Change-Id: I15cf89c240fddf91428a2e4cfdf1f1b21751e978
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>