Don't bind .gitignore file name to text content type
There is already a separate "Gitignore File" content type that the file
name is bound to (and that inherits from text content type). Binding it
to the text content type directly is not necessary, but rather confuses
end users, since they see the ".gitignore" name at the text content type
otherwise.

Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Change-Id: I3630e89c8b0ef968cf1dccc1b038527d0f5cd6cb
Bug: 580294