Show blame menu item in ruler only if not showing blame already
In the ruler context menu display "Show Revision Information" only
if the ruler actually can show revision information but doesn't do
so already.

For the user, this effectively toggles between the "Show Revision
Information" menu item and the "Revisions" sub-menu.

It is not necessary to display the "Show Revision Information" menu
item when revision information is already shown. EGit updates the
revision information automatically if and when it changes on a branch
switch or a commit.

Change-Id: I1bd5272e8ea8f505cbf41e1e31e628e5c4f0bbd5
Signed-off-by: Thomas Wolf <twolf@apache.org>