Fix symlink merge with stage 2 as input
Fix resolving a link/link merge conflict with stage 2 (previous HEAD)
as input. In such a conflict, the working tree file will be a symlink.
(In file/link or link/file conflicts, the working tree will contain the
file.)

Bring back the LocationEditableRevision. We cannot create a hidden
IResource for a broken symlink, and if the symlink has a valid target,
the IResource might update the target in the end. We have to handle
this specially.

Luckily, the reason for not using LocationEditableRevision anymore given
in commit 2824d518 (JDT logging exceptions and not initializing the
merge viewer fully) is irrelevant for symlinks, as those should not use
a Java document provider anyway.

Change-Id: I4d89dc489894b71b3289bd01b1c768f35bf0a238
Signed-off-by: Thomas Wolf <twolf@apache.org>