Use rename detection in some comparisons
Use rename detection for comparisons of a single file between index and
HEAD, or between a commit and an immediate parent.

Rename detection in general is an O(A * D) process, where A is the
number of added files and D is the number of deleted files. If copy
detection is done, it becomes O((A+M) * (D+M)), where M is the number
of modified files. But if one is interested in only a single path,
it can be done in O(D) or O(D+M).

Bug: 581030
Change-Id: Ie873a14a14c768a2697fd4b16517e0a9aaeb4a40
Signed-off-by: Thomas Wolf <twolf@apache.org>