[staging] "Compare With Each Other" for unstaged files
Implement comparing two selected unstaged files against each other. As
it turns out, org.eclipse.compare does not provide any way to do this
cleanly!? However, with access to one internal class it is possible.
Since EGit's CompareUtils already accessed two other internal classes
of org.eclipse.compare, this was deemed acceptable. But even then the
implementation is a bit of a hack.

For comparing files not in the Eclipse workspace, create hidden linked
resources. This is actually the same mechanism as the somewhat hidden
"Compare with Other Resource" command from org.eclipse.compare uses.
Unfortunately we can't use this command because it unconditionally
opens a dialog that is completely superfluous for the EGit use case.

Change-Id: Ic9db887a0ad40622f5043a2dfaf01b46e96db346
Signed-off-by: Thomas Wolf <twolf@apache.org>