Update Orbit reference in p2 repository
When using a different Orbit, *always* also update the repository
reference in org.eclipse.egit.repository/category.xml!

This was forgotten in commit ba081d68.

It's no big deal, but let's be consistent. We added this reference to
catch cases where the EGit p2 repository would lack a needed bundle, or
Orbit would have a newer version of some bundle, so that it could be
found automatically in the Orbit p2 repository, and people could still
easily install EGit from only the EGit repository. With an outdated
Orbit reference, that might not work, and users might need to add the
correct Orbit repo manually to their Eclipse update sites to be able to
install or upgrade EGit manually in an existing Eclipse.

Given that Eclipse platform is moving away from Orbit anyway and started
using maven bundles directly, the importance of this is diminishing
anyway.

Change-Id: Ice4068b4bddc41dcd0dbe1edabd325e904ee714f
Signed-off-by: Thomas Wolf <twolf@apache.org>