Avoid animated refresh of repository view
If the repository view is expanded to the branches node, changing the
hierarchical layout state will lead to an "animation" when
adding/removing the nodes underneath (easy to see with many branches
existing, at least on Windows). Disable the redrawing during the refresh
to make the full result of the refresh at once.

Change-Id: I9288b797003562da1820f999baf13a645378850a