Include OSGi annotations in the target platform
They are required by PDE, which we add to the Eclipse running the UI
tests. With Tycho 2.7.5, they are somehow present anyway (or 2.7.5
ignores them and starts the Eclipse anyway), but Tycho 4 will complain
and refuse to run the tests because it cannot fully resolve the target
platform for the running Eclipse.

Change-Id: I61fe4f7941a75e813d70affc71174a506cb3cfe4
Signed-off-by: Thomas Wolf <twolf@apache.org>